requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
pydantic==2.7.4
openai==1.40.0
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
